Tests: Integration tests add for customer type

+++ src/controllers/customerTypeController.ts
@@ -1,3 +1,4 @@
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/client";
 import type { Request, Response } from "express";
 import { prisma } from "../lib/prisma";
 
@@ -19,8 +20,18 @@ export const createCustomerType = async (req: Request, res: Response) => {
         const newCustomerType = await prisma.tipologiaCliente.create({
             data: { tipoCliente },
         });
+
         res.status(201).json(newCustomerType);
     } catch (error) {
+        if (
+            error instanceof PrismaClientKnownRequestError &&
+            error.code === "P2002"
+        ) {
+            return res.status(409).json({
+                error: "Cannot duplicate customer type value",
+            });
+        }
+
         res.status(500).json({ error: "Internal server error" });
     }
 }
@@ -32,13 +43,16 @@ export const deleteCustomerType = async (req: Request, res: Response) => {
         return res.status(400).json({error: "Missing parameter in the request: id"});
 
     try {
+        const customerType = await await prisma.tipologiaCliente.findUnique({
+            where: { id: id },
+        });
+        if (!customerType) {
+            return res.status(404).json({ error: "Customer type not found" });
+        }
+
         const deletedCustomerType = await prisma.tipologiaCliente.delete({
             where: { id: id },
         });
-
-        if (!deletedCustomerType) {
-            return res.status(404).json({ error: "Customer type not found" });
-        }
 
         res.json(deletedCustomerType);
     } catch (error) {
